requirements.txt
lyricsgenius
Pillow
